Add controller for device routes. #101
